ModuleTests/CodeShield/requirements.txt
httpx
pydantic
jsonschema
streamlit
plotly
pytest
pandas
python-dotenv
fastapi
uvicorn[standard]
python-multipart
